Changing get_cluskey() to return max 64

+++ offline/packages/trackbase_historic/SvtxTrackState.h
@@ -51,7 +51,7 @@ class SvtxTrackState : public PHObject
   virtual float get_error(unsigned int /*i*/, unsigned int /*j*/) const { return NAN; }
   virtual void set_error(unsigned int /*i*/, unsigned int /*j*/, float /*value*/) {}
 
-  virtual TrkrDefs::cluskey get_cluskey() const { return UINT32_MAX; }
+  virtual TrkrDefs::cluskey get_cluskey() const { return UINT64_MAX; }
   virtual void set_cluskey(TrkrDefs::cluskey) {}
 
   virtual std::string get_name() const { return ""; }
